fix tx数据导统一db，domain字段无效的问题
# message auto-generated for no-merge-commit merge:
merge cc into master

fix tx数据导统一db，domain字段无效的问题

Created-by: msprofDeveloper

Match-id-d531c98a78e0c9a1527c534402800f404dc94b7b

+++ analysis/csrc/viewer/database/finals/msproftx_processor.h
@@ -31,7 +31,7 @@ using MsprofTxExDataFormat = std::vector<std::tuple<uint32_t, std::string,
 
 // startNs, endNs, eventType, rangeId, category, message, globalTid, endGlobalTid, domainId, connectionId
 using ProcessedDataFormat = std::vector<std::tuple<uint64_t, uint64_t, uint16_t,
-        uint32_t, uint32_t, uint64_t, uint64_t, uint64_t, uint16_t, uint64_t>>;
+        uint32_t, uint32_t, uint64_t, uint64_t, uint64_t, uint64_t, uint64_t>>;
 
 public:
     MsprofTxProcessor() = default;
